Move more general functions into separate utils module; tweak some docstrings
